Automatically configure appropriate guest resolvers
By default, docker will use the hosts resolver list (read from resolv.conf) as a
list of resolvers for use on the guest containers. This is problematic if the
host resolver list includes the loopback address, which has a different meaning
within a container.

Automatically configure guest containers to use all the resolvers listed in
resolv.conf *except* for the loopback, and warn if this means there are no
resolvers in the list.